Fix Lab2

+++ src/itmo/java/basics/lab2/Calculator.java
@@ -13,16 +13,16 @@ public class Calculator {
         return operand1+operand2;
     }
 
-    public static int division (int operand1, int operand2){
-        return operand1/operand2;
+    public static double division (int operand1, int operand2){
+        return (double) operand1/operand2;
     }
 
     public static double division (double operand1, double operand2){
         return operand1/operand2;
     }
 
-    public static long division (long operand1, long operand2){
-        return operand1/operand2;
+    public static double division (long operand1, long operand2){
+        return (double) operand1/operand2;
     }
 
     public static int multiplication (int operand1, int operand2){
